book_summary_app/requirements.txt
openai>=1.10.0
langgraph>=0.0.20
langchain>=0.0.300
langchain-openai>=0.0.2
requests>=2.31.0
python-dotenv>=1.0.0 